+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - Fix facebook ranks sync exception when there is no post
   - Break line when post content is too long
   - Optimize block words filter
+  - Support markdown for terms of service and privacy policy
 
 - 1.0.1 (2016-06-08)
   - Fix incorrect form error message